add: seperate class file
