+++ product/src/main/kotlin/com/product/domain/product/application/ProductCommandServiceImpl.kt
@@ -1,9 +1,6 @@
 package com.product.domain.product.application
 
 import com.product.domain.product.dto.CreateProductDto
-import com.product.domain.product.dto.ProductDto
-import com.product.domain.product.dto.ProductsDto
-import com.product.domain.product.dto.SaleDto
 import com.product.domain.product.event.ProductCreatedEvent
 import com.product.global.kafka.consumer.dto.OrderReservedEvent
 import com.product.domain.product.event.ProductReservedEvent
@@ -17,17 +14,15 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.time.LocalDateTime
 import java.util.*
 
 @Service
-class ProductServiceImpl(
+class ProductCommandServiceImpl(
     private val applicationEventPublisher: ApplicationEventPublisher,
     private val categoryRepository: CategoryRepository,
     private val productRepository: ProductRepository,
-    private val mongoProductRepository: MongoProductRepository,
     private val orderApi: OrderApi
-) : ProductService {
+) : ProductCommandService {
 
     @DistributedLock(key = "'reserve_product_' + #event.products.![productId]")
     @Transactional
@@ -68,50 +63,6 @@ class ProductServiceImpl(
             product.add(it.quantity)
             productRepository.save(product)
         }
-    }
-
-    @Transactional(readOnly = true)
-    override fun queryProducts(categoryId: Long): ProductsDto {
-//        val products = productRepository.findByCategoryId(categoryId)
-//
-//        val productDto = products.map { product ->
-//            val orderCount = orderApi.queryOrderCount(product.id)
-//
-//            ProductDto(
-//                productId = product.id,
-//                categoryId = product.category.id,
-//                categoryName = product.category.name,
-//                name = product.name,
-//                originPrice = product.price,
-//                quantity = product.quantity,
-//                isSale = product.sale != null,
-//                orderCount = orderCount.count,
-//                totalOrderQuantity = orderCount.totalQuantity,
-//                saleInfo = product.sale?.let {
-//                    SaleDto(
-//                        salePercentage = it.percentage,
-//                        salePrice = product.price * it.percentage / 100,
-//                        saleStartDate = it.startDate,
-//                        saleEndDate = it.endDate,
-//                    )
-//                },
-//                createdDate = LocalDateTime.now(),
-//            )
-//        }
-
-//        return ProductsDto(
-//            count = productDto.size,
-//            categoryName = products[0].category.name,
-//            products = productDto
-//        )
-
-        TODO()
-    }
-
-    @Transactional(readOnly = true)
-    override fun queryProduct(productId: Long): ProductDocument {
-        return mongoProductRepository.findById(productId)
-            .orElseThrow { throw ProductException("Product with id ${productId} not found", HttpStatus.NOT_FOUND) }
     }
 
     @Transactional
