+++ product/src/main/kotlin/com/product/domain/product/application/ProductQueryService.kt
@@ -1,0 +1,7 @@
+package com.product.domain.product.application
+
+import com.product.domain.product.persistence.ProductDocument
+
+interface ProductQueryService {
+    fun queryProduct(productId: Long): ProductDocument
+}
