+++ product/src/main/kotlin/com/product/domain/product/presentation/ProductController.kt
@@ -1,6 +1,7 @@
 package com.product.domain.product.presentation
 
-import com.product.domain.product.application.ProductService
+import com.product.domain.product.application.ProductCommandService
+import com.product.domain.product.application.ProductQueryService
 import com.product.domain.product.dto.CreateProductDto
 import com.product.domain.product.persistence.ProductDocument
 import org.springframework.http.HttpStatus
@@ -10,24 +11,15 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/product")
 class ProductController(
-    private val productService: ProductService
+    private val productCommand: ProductCommandService,
+    private val productQuery: ProductQueryService,
 ) {
-
-    /*
-    @GetMapping
-    fun queryProducts(
-        @RequestParam categoryId: Long,
-    ): ResponseEntity<ProductsDto> {
-        val response = productService.queryProducts(categoryId)
-        return ResponseEntity.ok(response)
-    }
-    */
 
     @PostMapping
     fun createProduct(
         @RequestBody dto: CreateProductDto
     ): ResponseEntity<Map<String, Long>> {
-        val response = productService.createProduct(dto)
+        val response = productCommand.createProduct(dto)
         return ResponseEntity.status(HttpStatus.CREATED.value()).body(mapOf("productId" to response))
     }
 
@@ -35,14 +27,8 @@ class ProductController(
     fun queryProduct(
         @PathVariable("product_id") productId: Long,
     ): ResponseEntity<ProductDocument> {
-        val response = productService.queryProduct(productId)
+        val response = productQuery.queryProduct(productId)
         return ResponseEntity.ok(response)
     }
-
-    // product:
-    // Hibernate: select p1_0.id,p1_0.category_id,c1_0.id,c1_0.name,p1_0.created_date,p1_0.name,p1_0.price,p1_0.quantity,s1_0.id,s1_0.end_date,s1_0.name,s1_0.percentage,s1_0.start_date from tbl_product p1_0 join tbl_category c1_0 on c1_0.id=p1_0.category_id left join tbl_sale s1_0 on s1_0.id=p1_0.sale_id where p1_0.id=?
-
-    // order:
-    // Hibernate: select op1_0.id,op1_0.order_id,op1_0.product_id,op1_0.quantity from tbl_order o1_0 join tbl_order_product op1_0 on op1_0.order_id=o1_0.order_id where op1_0.product_id=?
 
 }
